test: fix resource lock during snapshot auto update tests
